extract candidates loader
